add backend

+++ internship-server/src/user/user.service.ts
@@ -1,18 +1,24 @@
 import { Injectable } from '@nestjs/common';
-import { InjectInMemoryDBService, InMemoryDBService } from '@nestjs-addons/in-memory-db';
+import {
+  InjectInMemoryDBService,
+  InMemoryDBService,
+} from '@nestjs-addons/in-memory-db';
 import { User } from './user';
 import { UserDto } from './user.dto';
 
 @Injectable()
 export class UserService {
-  constructor(@InjectInMemoryDBService('user') private readonly database: InMemoryDBService<User>) {}
+  constructor(
+    @InjectInMemoryDBService('user')
+    private readonly database: InMemoryDBService<User>,
+  ) {}
 
   getAllUsers(): User[] {
     return this.database.getAll();
   }
 
   async getUserByEmail(email: string): Promise<User | undefined> {
-    const users = this.database.query(user => user.email === email);
+    const users = this.database.query((user) => user.email === email);
     return users.length > 0 ? users[0] : undefined;
   }
 
@@ -28,14 +34,17 @@ export class UserService {
   async deleteUser(email: string) {
     const user = await this.getUserByEmail(email);
 
-    if(user)
-      this.database.delete(user.id);
+    if (user) this.database.delete(user.id);
   }
 
   dtoToUser(userDto: UserDto) {
     let user = new User();
     user.email = userDto.email;
     user.password = userDto.password;
+    user.company = userDto.company;
+    user.firstName = userDto.firstName;
+    user.lastName = userDto.lastName;
+    user.roleId = userDto.roleId;
 
     return user;
   }
